trading_bot/requirements.txt
# Core Dependencies
ccxt>=4.0.0  # Unified crypto exchange API
python-dotenv>=1.0.0
requests>=2.31.0

# Web Framework
fastapi>=0.95.0
uvicorn[standard]>=0.21.1
pydantic>=1.10.7

# Data Processing
pandas>=2.0.0
numpy>=1.24.0
python-dateutil>=2.8.2  # Date parsing and manipulation
pytz>=2023.3  # Timezone support

# Logging
structlog>=22.3.0  # Structured logging

# Performance
orjson>=3.8.0  # Faster JSON parsing

# Security
cryptography>=40.0.0  # For secure data handling
python-jose[cryptography]>=3.3.0

# Testing
pytest>=7.3.1
pytest-asyncio>=0.21.0
pytest-cov>=4.0.0

# Monitoring
prometheus-client>=0.16.0  # For metrics collection

# Development
black>=23.3.0
isort>=5.12.0
mypy>=1.0.0
flake8>=6.0.0
